Form switch statement for operator press listener
In the process of forming the switch statement and
using the right code in the right places. I am currently trying
to use the equalssign event listener code for when the button is pressed the second time.
I have identified that the button being pressed a second time via a switch statement assessing
whether or not any of the operators are in the display as of pressing a new operator button.